adding work duplication checking and skipping, converting config.json to use all subregions by default.

+++ utils/mergeMbtiles.js
@@ -14,13 +14,54 @@ export const mergeMbtiles = async (subRegions, keepSubRegionMbtiles) => {
   }
 
   // Set up the base planet.mbtiles file to merge all other files into. Overwrite existing file.
-  fs.writeFileSync(`${pwd}/mbtiles/merged/planet.mbtiles`, "");
+  if (!fs.existsSync(`${pwd}/mbtiles/merged/planet.mbtiles`)) {
+    fs.writeFileSync(`${pwd}/mbtiles/merged/planet.mbtiles`, "");
+  }
+
+  const mbtilesPreFileSize = fs.statSync(
+    `${pwd}/mbtiles/merged/planet.mbtiles`
+  ).size;
 
   execute(
     "tile-join",
     ["-o", "merged/planet.mbtiles", ...mbtilesToMerge],
     `${pwd}/mbtiles`
   );
+
+  const mbtilesPostFileSize = fs.statSync(
+    `${pwd}/mbtiles/merged/planet.mbtiles`
+  ).size;
+
+  console.log(`${mbtilesPreFileSize} => ${mbtilesPostFileSize}`);
+  if (mbtilesPreFileSize < mbtilesPostFileSize) {
+    console.log(
+      "mbtiles file size increased. The process appears to have completed successfully."
+    );
+    fs.appendFileSync(
+      `${pwd}/mbtiles/REPORT.txt`,
+      `
+    --------------------------------------------------
+    SUCCESSFULLY MERGED MBTILES FILES - see above for any tilesets which failed to be created, and therefore failed to be merged.
+
+    Remember to check the files before production.
+    --------------------------------------------------
+    `
+    );
+  } else {
+    console.log(
+      "mbtiles file size did not increase. The process appears to have failed."
+    );
+    fs.appendFileSync(
+      `${pwd}/mbtiles/REPORT.txt`,
+      `
+      --------------------------------------------------
+      FAILED TO MERGE MBTILES FILES - see above for any tilesets which failed to be created. 
+      
+      If you are seeing this message it is becuase either no (new) mbtiles were provided to be merged (i.e. there were no sub regions), or the joining process failed for some reason.
+      --------------------------------------------------
+      `
+    );
+  }
 
   if (!keepSubRegionMbtiles) {
     // Delete the sub region mbtiles files if required.
